Fix display filter
The old implementation did not return the <result> root element. We
want to keep it so we have to modify our manipulation process to
rely less on Fluid and more on Path

The new XPath removes items which do not have a child skærm item with
the required name.