Implement support for sorting of events
Events are now sorted by start date as that affects the order they
are displayed in in OS2Display.

Since this only applies for events this is one more break from the
original content agnostic approach of the proxy. To handle this means
yet another special case handling - this type dependent on both
path (to avoid sorting other types of content) and content (we only
support sorting XML).

Add a test to show that it works.